mise a jour de la classe EntrepriseDao pour prendre en compte la connexion à la db par la classe ConnexionDatabase. Insertion des requete QUERY preconstruite dans u fichier properties

+++ src/com/cdb/entities/Entreprise.java
@@ -5,7 +5,7 @@ public class Entreprise {
 	private long id;
 	private String name;
 
-	public Entreprise(int id, String name) {
+	public Entreprise(long id, String name) {
 		
 		super();
 		this.id = id;
